split out snapshot and historical to their own files & tests
historical: remove fields ref

historical.spec: remove snapshot assertion (to merge)
