+++ lib/snapshot.spec.js
@@ -67,42 +67,21 @@ function assertValidSnapshotResult(result) {
   result.should.include.keys('symbol', 'name');
 }
 
-describe('index', () => {
+describe('snapshot', () => {
 
-  describe('historical', () => {
+  it('correctly transforms current Yahoo response (2017-05-21)', async () => {
 
-    it('correctly transforms current Yahoo response (2017-05-21)', async () => {
-
-      const historical = proxyquire('./index', stubbedFor('historyCsv')).historical;
-      const result = await historical({
-        // Note, these aren't actually used in this test - data from a fixture
-        symbol: 'AAPL',
-        from: '2012-01-01'
-      });
-
-      assertValidHistoricalResult(result);
-
+    const snapshot = proxyquire('./index', stubbedFor('quoteJson')).snapshot;
+    const result = await snapshot({
+      // Note, these aren't actually used in this test - data from a fixture
+      symbol: 'TSLA',
+      fields: ['s', 'n', 'd1', 'l1', 'y', 'r']
     });
 
-  });
-
-  describe('snapshot', () => {
-
-    it('correctly transforms current Yahoo response (2017-05-21)', async () => {
-
-      const snapshot = proxyquire('./index', stubbedFor('quoteJson')).snapshot;
-      const result = await snapshot({
-        // Note, these aren't actually used in this test - data from a fixture
-        symbol: 'TSLA',
-        fields: ['s', 'n', 'd1', 'l1', 'y', 'r']
-      });
-
-      assertValidSnapshotResult(result);
-
-    });
+    assertValidSnapshotResult(result);
 
   });
 
 });
 
-export { assertValidSnapshotResult, assertValidHistoricalResult };
+export { assertValidSnapshotResult };
